Week 3 team completed

+++ controllers/index.js
@@ -12,25 +12,35 @@ async function connectDB() {
     }
 }   
 
+const { ObjectId } = require("mongodb");
+const NotFoundError = require('../errors/notFoundError'); 
+const DatabaseError = require('../errors/databaseError');
+const UserDataError = require('../errors/userDataError'); 
 
-async function retrieve(req, res) {
-    await connectDB();
-    const { ObjectId } = require("mongodb"); 
-    const collection = db.collection("current"); 
-    const { id } = req.params; 
-    console.log("id: ", id)
+async function retrieve(req, res, next) {
     try {
+    await connectDB();
+    const collection = db.collection("current"); 
+        const { id } = req.params;
         if (id) {
-            const data = await collection.findOne({_id : new ObjectId(id) }); 
-            res.json(data); 
-        } else {
-            const data = await collection.find({}).toArray(); 
-            res.json(data); 
+            
+            if (ObjectId.isValid(id)) {
+                const data = await collection.findOne({ _id: new ObjectId(id) });
+                if (!data) {
+                    throw new NotFoundError(`User ${id} does not exist in database`);
+                };
+                res.json(data);
+            } else {
+                throw new NotFoundError("Please ensure you entered a valid Mongo DB Id.");
+            }
         }
-        
-    } catch (error) {
-        console.error(error); 
-        res.status(500).send("Error fetching Data");
+                const data = await collection.find({}).toArray(); 
+        if (data.length === 0) {
+                    throw new NotFoundError("No contacts found in the database.")
+                };
+            res.json(data); 
+        } catch (error) {
+        next(error); 
     }
 }
     
@@ -41,7 +51,6 @@ async function createContact(req, res) {
     const email = req.body.email || null;
     const favoriteColor = req.body.favoriteColor || null;
     const birthday = req.body.birthday || null; 
-    console.log(firstName, lastName, email, favoriteColor, birthday)
         
         
     if (firstName && lastName && email && favoriteColor && birthday) {
@@ -53,7 +62,6 @@ async function createContact(req, res) {
             birthday : req.body.birthday 
         }
         
-        console.log("Entry: " ,entry); 
         await connectDB();
         const collection = db.collection("current");
         try {
@@ -61,10 +69,10 @@ async function createContact(req, res) {
             res.json(result);
         } catch (error) {
             console.error(error);
-            res.status(500).send("Data not saved. There was an uploading error.  Try again. ")
+            throw new DatabaseError("Data not saved. There was an uploading error.  Try again. ")
         }
     }else {
-        res.send("You must provide all information for the contact.  Please include firstName, lastName, email, favoriteColor, and birthday. ")
+        throw new UserDataError("You must provide all information for the contact.  Please include firstName, lastName, email, favoriteColor, and birthday. ")
     }
     
 }
@@ -73,25 +81,28 @@ async function createContact(req, res) {
 async function updateContact(req, res) {
     const updates = { ...req.body }
     const { ObjectId } = require("mongodb"); 
-        await connectDB();
-        const collection = db.collection("current"); 
-    const record = await collection.findOne({_id: new ObjectId(updates.id)});
-    delete updates.id; 
-    console.log("Record: ", record)
+    await connectDB();
+    const collection = db.collection("current"); 
+    if (ObjectId.isValid(updates.id)) {
+        
+        
+        const record = await collection.findOne({_id: new ObjectId(updates.id)});
+        delete updates.id; 
+        console.log("Record: ", record)
         if (record) {
             try {
                 const result = await collection.updateOne({ _id: record._id },
                     { $set: updates })
-                console.log("reached 1")
-                ""; 
-                console.log("reached 2")
-                res.json(result); 
-            } catch (error) {
-                console.error(error); 
-                res.status(500).send("Database error. Record not updated.  Please try again.")
+                    res.json(result); 
+                } catch (error) {
+                    console.error(error); 
+                    throw new DatabaseError("Database error. Record not updated.  Please try again.")
                 }
-        } else {
-            res.status(404).send("Record not found.  Try another ID.")
+            } else {
+                throw new NotFoundError("Record not found.  Try another ID.")
+            }
+    } else {
+        throw new NotFoundError("Please enter a valid MondoDB Id.")
         }
 }
 
@@ -99,22 +110,22 @@ async function deleteContact(req, res) {
     const { ObjectId } = require("mongodb");
     await connectDB();
     const collection = db.collection("current");
-    const record = await collection.findOne({ _id: new ObjectId(req.params.id) });
-    console.log("Record: ", record)
-    if (record) {
-        try {
-            const result = collection.deleteOne({ _id: record._id });
-            res.send("Contact Removed"); 
-            
-        } catch (error) {
-            console.error(error);
-            res
-            .status(500)
-            .send("Database error. Record not deleted.  Please try again.");
+    if (ObjectId.isValid(req.params.id)) {
+        
+        
+        const record = await collection.findOne({ _id: new ObjectId(req.params.id) });
+        if (record) {
+            try {
+                const result = collection.deleteOne({ _id: record._id });
+                res.send("Contact Removed"); 
+                
+            } catch (error) {
+                throw new DatabaseError("Database error. Record not deleted.  Please try again.");
+            }
+        } else {
+            throw new NotFoundError("Record not found.  Try another ID.");
         }
-    } else {
-        res.status(404).send("Record not found.  Try another ID.");
-    }
+    } else throw new NotFoundError("Please enter a valid MongoDB Id.")
 }
